+++ examples/03_connectivity/plot_extract_regions_dictlearning_maps.py
@@ -20,8 +20,8 @@ for more details.
 
     The use of the attribute `components_img_` from dictionary learning
     estimator is implemented from version 0.4.1. For older versions,
-    unmask the attribute `components_` to get the components image using
-    attribute `masker_` embedded in estimator.
+    unmask the deprecated attribute `components_` to get the components
+    image using attribute `masker_` embedded in estimator.
     See the :ref:`section Inverse transform: unmasking data <unmasking_step>`.
 """