+++ nilearn/decomposition/canica.py
@@ -116,11 +116,13 @@ class CanICA(MultiPCA):
         Masked ICA components extracted from the input images. They can be
         unmasked thanks to the `masker_` attribute.
 
+        Deprecated since version 0.4.1. Use `components_img_` instead.
+
     `components_img_` : 4D Nifti image
         4D image giving the extracted ICA components. Each 3D image is a
         component.
 
-        .. versionadded:: 0.4.1
+        New in version 0.4.1.
 
     `masker_` : instance of MultiNiftiMasker
         Masker used to filter and mask data as first step. If an instance of