+++ nilearn/decomposition/tests/test_canica.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from numpy.testing import assert_array_almost_equal
-from nose.tools import assert_true, assert_raises, assert_not_equal
+from nose.tools import assert_true, assert_raises
 import nibabel
 
 from nilearn._utils.testing import (assert_less_equal,